requirements.txt
alembic==1.12.0
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
distlib==0.3.7
filelock==3.12.3
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.5
flask-mysql-connector==1.1.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.1.1
numpy==1.25.2
pandas==2.1.0
platformdirs==3.10.0
protobuf==4.21.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.4
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
